Clean up ticket title and component for support error notice
